fix: wrong keybinding variable syntax
Bash does not expand variables inside double-quoted strings in bind commands, preventing the shortcut variable from being recognized.

To ensure variable expansion, the shortcut variable is enclosed in a combination of single and double quotes, effectively breaking out of the initial double-quoted context to allow the variable to expand, and then re-entering a double-quoted context for the bind command.